Reworking functional tests for pre and post Jenkins calls
