--- src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/controllers/FunctionalTest.java
@@ -1,41 +1,0 @@
-package uk.gov.hmcts.reform.sscs.controllers;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
-import uk.gov.hmcts.reform.sscs.idam.IdamService;
-import uk.gov.hmcts.reform.sscs.services.sftp.SftpChannelAdapter;
-
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-@AutoConfigureMockMvc
-public class FunctionalTest {
-
-    @MockBean
-    private SftpChannelAdapter channelAdapter;
-    @MockBean
-    private CcdService ccdService;
-    @MockBean
-    private IdamService idamService;
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    public void shouldReturn200WhenSendingRequestToController() throws Exception {
-        mockMvc.perform(get("/functional-test"))
-            .andDo(print())
-            .andExpect(status().isOk());
-    }
-
-}
