+++ src/e2e/java/uk/gov/hmcts/reform/sscs/functional/predeploy/ProcessCaseFileTest.java
@@ -1,0 +1,52 @@
+package uk.gov.hmcts.reform.sscs.functional.predeploy;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+import lombok.extern.slf4j.Slf4j;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
+import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
+import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
+import uk.gov.hmcts.reform.sscs.services.CaseLoaderService;
+import uk.gov.hmcts.reform.sscs.services.sftp.SftpChannelAdapter;
+
+@RunWith(SpringRunner.class)
+@TestPropertySource(locations = "classpath:config/application_e2e.yaml")
+@SpringBootTest
+@Slf4j
+public class ProcessCaseFileTest {
+
+    @Autowired
+    private SftpChannelAdapter sftpChannelAdapter;
+    @Autowired
+    private CcdService ccdService;
+    @Autowired
+    private CaseLoaderService caseLoaderService;
+
+    private String ccdCaseId;
+    private IdamTokens idamTokens;
+
+    @SuppressWarnings("unchecked")
+    @Test
+    public void processCaseFileAndVerifyCcd() {
+
+        caseLoaderService.process();
+
+        SscsCaseDetails updatedCcdCase = ccdService.getByCaseId(Long.parseLong(ccdCaseId), idamTokens);
+        assertNotNull(updatedCcdCase);
+
+        SscsCaseData updatedCcdCaseData = updatedCcdCase.getData();
+
+        assertEquals("XYZ", updatedCcdCaseData.getAppeal().getAppellant().getName().getFirstName());
+        assertEquals(3, updatedCcdCaseData.getEvents().size());
+        assertEquals("appealCreated", updatedCcdCase.getState());
+    }
+}
