--- src/main/java/uk/gov/hmcts/reform/sscs/controller/test/smoke/Smoke.java
@@ -1,35 +1,0 @@
-package uk.gov.hmcts.reform.sscs.controller.test.smoke;
-
-import com.google.common.collect.ImmutableMap;
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
-import uk.gov.hmcts.reform.sscs.idam.IdamService;
-import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
-
-@Controller
-public class Smoke {
-
-    @Autowired
-    private CcdService ccdService;
-    @Autowired
-    private IdamService idamService;
-
-    @GetMapping("/smoke-test")
-    @ResponseBody
-    public List<SscsCaseDetails> smoke() {
-        String oauth2Token = idamService.getIdamOauth2Token();
-        IdamTokens idamTokens = IdamTokens.builder()
-            .idamOauth2Token(oauth2Token)
-            .serviceAuthorization(idamService.generateServiceAuthorization())
-            .userId(idamService.getUserId(oauth2Token))
-            .build();
-        return ccdService.findCaseBy(
-            ImmutableMap.of("case.caseReference", "SC068/18/01217"), idamTokens);
-    }
-
-}
